Solución P05 ASM corregido a 16 bits

+++ Clase_06_adm_c16/Core/Inc/asm_func.h
@@ -25,8 +25,8 @@ void asm_productoEscalar12(	uint16_t *vectorIn,
 						uint16_t escalar);
 
 
-void asm_filtroVentana10(	uint32_t * vectorIn,
-						uint32_t * vectorOut,
+void asm_filtroVentana10(	uint16_t * vectorIn,
+						uint16_t * vectorOut,
 						uint32_t longitudVectorIn);
 
 void asm_invertir(		uint16_t *vector,
